third_party/WebKit/LayoutTests/media/video-currentTime-before-have-metadata.html

<!DOCTYPE html>
<html>
    <head>
        <script src=media-file.js></script>
        <script src=video-test.js></script>
        <script>
            function onWindowLoad(e)
            {
                video = document.getElementById('video');

                video.src = findMediaFile("video", "content/test");
                testExpected("video.currentTime", 0);
                video.currentTime = 1;
                testExpected("video.currentTime", 1);

                waitForEvent('loadedmetadata', function()
                {
                    testExpected("video.currentTime", 1);
                });

                waitForEventAndEnd('seeked');
            }

            window.addEventListener('load', onWindowLoad, false);
        </script>
    </head>
    <body>
        <video controls id="video"></video>
        <p>Test currentTime values when setting while HAVE_NOTHING.</p>
        <br/>
    </body>
</html>
